Separate generic pipeline methods into a separate module
