+++ src/toast/CMakeLists.txt
@@ -137,3 +137,4 @@ add_subdirectory(tests)
 add_subdirectory(tod)
 add_subdirectory(fod)
 add_subdirectory(map)
+add_subdirectory(pipeline_tools)
