+++ src/toast/utils.py
@@ -2,6 +2,8 @@
 # All rights reserved.  Use of this source code is governed by
 # a BSD-style license that can be found in the LICENSE file.
 
+import gc
+
 import numpy as np
 
 from ._libtoast import Environment, Timer, GlobalTimers, Logger
@@ -32,6 +34,93 @@ from ._libtoast import (
     vfast_log,
     vfast_erfinv,
 )
+
+
+try:
+    import psutil
+
+    def memreport(comm=None, msg=""):
+        """ Gather and report the amount of allocated, free and swapped system memory
+        """
+        if psutil is None:
+            return
+        vmem = psutil.virtual_memory()._asdict()
+        gc.collect()
+        vmem2 = psutil.virtual_memory()._asdict()
+        memstr = "Memory usage {}\n".format(msg)
+        for key, value in vmem.items():
+            value2 = vmem2[key]
+            if comm is None:
+                vlist = [value]
+                vlist2 = [value2]
+            else:
+                vlist = comm.gather(value)
+                vlist2 = comm.gather(value2)
+            if comm is None or comm.rank == 0:
+                vlist = np.array(vlist, dtype=np.float64)
+                vlist2 = np.array(vlist2, dtype=np.float64)
+                if key != "percent":
+                    # From bytes to better units
+                    if np.amax(vlist) < 2 ** 20:
+                        vlist /= 2 ** 10
+                        vlist2 /= 2 ** 10
+                        unit = "kB"
+                    elif np.amax(vlist) < 2 ** 30:
+                        vlist /= 2 ** 20
+                        vlist2 /= 2 ** 20
+                        unit = "MB"
+                    else:
+                        vlist /= 2 ** 30
+                        vlist2 /= 2 ** 30
+                        unit = "GB"
+                else:
+                    unit = "% "
+                if comm is None or comm.size == 1:
+                    memstr += "{:>12} : {:8.3f} {}\n".format(key, vlist[0], unit)
+                    if np.abs(vlist2[0] - vlist[0]) / vlist[0] > 1e-3:
+                        memstr += "{:>12} : {:8.3f} {} (after GC)\n".format(
+                            key, vlist2[0], unit
+                        )
+                else:
+                    med1 = np.median(vlist)
+                    memstr += (
+                        "{:>12} : {:8.3f} {}  < {:8.3f} +- {:8.3f} {}  "
+                        "< {:8.3f} {}\n".format(
+                            key,
+                            np.amin(vlist),
+                            unit,
+                            med1,
+                            np.std(vlist),
+                            unit,
+                            np.amax(vlist),
+                            unit,
+                        )
+                    )
+                    med2 = np.median(vlist2)
+                    if np.abs(med2 - med1) / med1 > 1e-3:
+                        memstr += (
+                            "{:>12} : {:8.3f} {}  < {:8.3f} +- {:8.3f} {}  "
+                            "< {:8.3f} {} (after GC)\n".format(
+                                key,
+                                np.amin(vlist2),
+                                unit,
+                                med2,
+                                np.std(vlist2),
+                                unit,
+                                np.amax(vlist2),
+                                unit,
+                            )
+                        )
+        if comm is None or comm.rank == 0:
+            print(memstr, flush=True)
+        if comm is not None:
+            comm.Barrier()
+        return
+
+except ImportError:
+
+    def memreport(comm=None, msg=""):
+        return
 
 
 def object_ndim(x):
